// frontend/profdiff-web-application/src/app/models/dto/renderedTreeNode.ts
/**
 * Profdiff web-app API
 *
 * Contact: jirkadanek@icloud.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RenderedTreeNodeNodeContent } from './renderedTreeNodeNodeContent';
import { RenderedTreeNodeMarker } from './renderedTreeNodeMarker';

/**
 * The structured representation of the requested compilation tree.
 */
export interface RenderedTreeNode {
  /**
   * The comparison status of this node.
   */
  marker: RenderedTreeNodeMarker;
  /**
   * The primary content to render (represents the only state for IDENTITY, INSERT, DELETE, and the NEW state for RELABEL).
   */
  content: RenderedTreeNodeNodeContent;
  /**
   * List of child nodes (if no children, then an empty array passed).
   */
  children: Array<RenderedTreeNode>;
}
export namespace RenderedTreeNode {}