// frontend/profdiff-web-application/src/app/models/dto/renderedTreeNodeMarker.ts
/**
 * Profdiff web-app API
 *
 * Contact: jirkadanek@icloud.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * The comparison status of this node.
 */
export const RenderedTreeNodeMarker = {
  Identity: 'IDENTITY',
  Insert: 'INSERT',
  Delete: 'DELETE',
  Relabel: 'RELABEL',
  Info: 'INFO',
  Neutral: 'NEUTRAL',
} as const;
export type RenderedTreeNodeMarker =
  (typeof RenderedTreeNodeMarker)[keyof typeof RenderedTreeNodeMarker];